+++ blockcypher/src/config.ts
@@ -1,8 +1,8 @@
 import { Requester } from '@chainlink/external-adapter'
-import { Config } from '@chainlink/types'
+import { ConfigFactory } from '@chainlink/types'
 
 export const DEFAULT_DATA_PATH = 'result'
 export const DEFAULT_CONFIRMATIONS = 6
 export const DEFAULT_ENDPOINT = 'balance'
 
-export const makeConfig = (prefix?: string): Config => Requester.getDefaultConfig(prefix)
+export const makeConfig: ConfigFactory = (prefix?) => Requester.getDefaultConfig(prefix)