+++ btc.com/test/balance.test.ts
@@ -1,4 +1,4 @@
 import { makeExecute } from '../src/adapter'
 import { shouldBehaveLikeBalanceAdapter } from '@chainlink/adapter-test-helpers'
 
-shouldBehaveLikeBalanceAdapter(makeExecute(), ['bitcoinmainnet'])
+shouldBehaveLikeBalanceAdapter(makeExecute(), ['bitcoin_mainnet'])