+++ cryptoapis/test/difficulty.test.ts
@@ -1,10 +1,12 @@
 import { assert } from 'chai'
-import { Requester, assertSuccess, assertError } from '@chainlink/external-adapter'
+import { Requester } from '@chainlink/external-adapter'
+import { assertSuccess, assertError } from '@chainlink/adapter-test-helpers'
 import { AdapterRequest } from '@chainlink/types'
-import { executeWithDefaults } from '../src/adapter'
+import { makeExecute } from '../src/adapter'
 
 describe('difficulty endpoint', () => {
   const jobID = '1'
+  const execute = makeExecute()
 
   context('successful calls @integration', () => {
     const requests = [
@@ -26,7 +28,7 @@ describe('difficulty endpoint', () => {
 
     requests.forEach((req) => {
       it(`${req.name}`, async () => {
-        const data = await executeWithDefaults(req.testData as AdapterRequest)
+        const data = await execute(req.testData as AdapterRequest)
         assertSuccess({ expected: 200, actual: data.statusCode }, data, jobID)
         assert.isAbove(data.result, 0)
         assert.isAbove(data.data.result, 0)
@@ -51,7 +53,7 @@ describe('difficulty endpoint', () => {
     requests.forEach((req) => {
       it(`${req.name}`, async () => {
         try {
-          await executeWithDefaults(req.testData as AdapterRequest)
+          await execute(req.testData as AdapterRequest)
         } catch (error) {
           const errorResp = Requester.errored(jobID, error)
           assertError({ expected: 400, actual: errorResp.statusCode }, errorResp, jobID)