+++ example/src/endpoint/example.ts
@@ -1,5 +1,5 @@
 import { Requester, Validator } from '@chainlink/external-adapter'
-import { AdapterRequest, Config } from '@chainlink/types'
+import { ExecuteWithConfig } from '@chainlink/types'
 
 export const Name = 'example'
 
@@ -10,8 +10,8 @@ const customParams = {
   quote: ['quote', 'to', 'market'],
 }
 
-export const execute = async (config: Config, request: AdapterRequest) => {
-  const validator = new Validator(request, customParams)
+export const execute: ExecuteWithConfig = async (input, config) => {
+  const validator = new Validator(input, customParams)
   if (validator.error) throw validator.error
 
   const base = validator.validated.data.base