+++ sochain/src/adapter.ts
@@ -1,6 +1,6 @@
-import { Requester, Validator } from '@chainlink/external-adapter'
-import { Execute, ExecuteWithDefaults } from '@chainlink/types'
-import { Config, getConfig, logConfig, DEFAULT_ENDPOINT } from './config'
+import { Requester, Validator, AdapterError } from '@chainlink/external-adapter'
+import { ExecuteFactory, ExecuteWithConfig, Config } from '@chainlink/types'
+import { makeConfig, DEFAULT_ENDPOINT } from './config'
 import { balance } from './endpoint'
 
 const inputParams = {
@@ -8,11 +8,11 @@ const inputParams = {
 }
 
 // Export function to integrate with Chainlink node
-export const execute: Execute = async (request, config: Config) => {
+export const execute: ExecuteWithConfig = async (request, config) => {
   const validator = new Validator(request, inputParams)
   if (validator.error) throw validator.error
 
-  logConfig(config)
+  Requester.logConfig(config)
 
   const jobRunID = validator.validated.id
   const endpoint = validator.validated.data.endpoint || DEFAULT_ENDPOINT
@@ -24,13 +24,17 @@ export const execute: Execute = async (request, config: Config) => {
       break
     }
     default: {
-      throw Error(`Endpoint ${endpoint} not supported.`)
+      throw new AdapterError({
+        jobRunID,
+        message: `Endpoint ${endpoint} not supported.`,
+        statusCode: 400,
+      })
     }
   }
 
   return Requester.success(jobRunID, response)
 }
 
-// Export function to integrate with Chainlink node
-export const executeWithDefaults: ExecuteWithDefaults = async (request) =>
-  execute(request, getConfig())
+export const makeExecute: ExecuteFactory<Config> = (config) => {
+  return async (request) => execute(request, config || makeConfig())
+}